requirements.txt
pyautogui==0.9.52
pystray==0.16.1
Pillow==8.3.2
